refactor(datastorage): split server.go into construction/routes/lifecycle (Wave 3)
Split pkg/datastorage/server/server.go (1,111 lines) into server.go (trimmed:
Server struct + ServerDeps), server_construction.go (NewServer + dependency
wiring), server_routes.go (Handler() + chi route table), and
server_lifecycle.go (Start/Shutdown + graceful-shutdown steps).

Pure code motion, no behavior change, no exported-API change. Removed an
unused pkg/cert import from server_construction.go left over from moving
loadSigningCertificate to server_lifecycle.go. See audit report
GO-ANTIPATTERN-AUDIT-2026-07-01 §7k-3.
